pyproject.toml
[tool.poetry]
name = "reposcan"
version = "0.1.0"
description = "Example GitHub app to gather basic repo data"
authors = ["Eric Rasmussen"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
certifi = "2021.10.8"
cffi = "1.15.0"
charset-normalizer = "2.0.10"
click = "8.0.3"
colorama = "0.4.4"
cryptography = "36.0.1"
deprecated = "1.2.13"
flask = "2.0.2"
idna = "3.3"
itsdangerous = "2.0.1"
jinja2 = "3.0.3"
markupsafe = "2.0.1"
pycparser = "2.21"
pygithub = "1.55"
pyjwt = "2.3.0"
pynacl = "1.4.0"
requests = "2.27.1"
six = "1.16.0"
urllib3 = "1.26.8"
werkzeug = "2.0.2"
wrapt = "1.13.3"

[tool.poetry.scripts]
ghlisten = 'reposcan.main:main'

[tool.poetry.group.dev.dependencies]
mypy = "^1.0.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
